requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
google-genai
python-dotenv
openpyxl